linux/log2.h: Fix 64bit calculations in roundup/down_pow_two()
Some users need to make sure their rounding function accepts and returns
64bit long variables regardless of the architecture. Sadly
roundup/rounddown_pow_two() takes and returns unsigned longs. It turns
out ilog2() already handles 32/64bit calculations properly, and being
the building block to the round functions we can rework them as a
wrapper around it.

Suggested-by: Robin Murphy <robin.murphy@arm.com>

+++ drivers/usb/gadget/udc/atmel_usba_udc.c
@@ -1076,7 +1076,7 @@ found_ep:
 		case USB_ENDPOINT_XFER_INT:
 			if (desc->wMaxPacketSize == 0)
 				ep->fifo_size =
-				    roundup_pow_of_two(_ep->maxpacket_limit);
+				    roundup_pow_of_two((u16)_ep->maxpacket_limit);
 			else
 				ep->fifo_size =
 				    roundup_pow_of_two(le16_to_cpu(desc->wMaxPacketSize));
